requirements.txt
flask
flask-cors
pdfplumber
beautifulsoup4
requests
itsdangerous
python-dotenv
